api/requirements.txt
# Core Dependencies
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

# LangChain Core
langchain==0.3.12
langchain-community==0.3.12
langchain-text-splitters==0.3.3
langgraph==0.2.60

# LLM Integrations
langchain-openai==0.2.13
langchain-google-genai==2.0.0

# Vector Store & Database
supabase==2.11.0
vecs==0.4.4
langchain-postgres==0.1.1

# Document Processing
beautifulsoup4==4.12.3
pypdf==4.3.1
openpyxl==3.1.5
python-docx==1.1.2
chardet==5.2.0

# FastAPI (alternative to Flask)
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2

# Utilities
numpy==1.26.4
pandas==2.2.2
tiktoken==0.8.0
python-multipart==0.0.16